+++ lib/osm.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/bounds.dart';
 export 'src/element.dart';
+export 'src/filter.dart';
 export 'src/pbf/exception.dart';
 export 'src/pbf/file.dart';
 export 'src/pbf/header.dart';